오류 수정및 브랜드,어코드 crud 생성

+++ academyfinal/src/main/java/com/hiki/academyfinal/dto/BrandDto.java
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 
 @Data @AllArgsConstructor @NoArgsConstructor @Builder
 public class BrandDto {
-	private Integer brandNo;
+	private long brandNo;
 	private String brandName;
 }
